added pipeline functions to main package, bumped version
