+++ README.md
@@ -3,11 +3,11 @@
 <h3 align="center">Code for preprocessing and building models with local field potentials</h3>
 
 
-See `feature_pipeline.py` and `prediction_pipeline.py` for usage.
-
 <p align="center">
 <img align="middle" src="example_cpsd.gif" width="600" height="600" />
 </p>
+
+See `script.py` for usage.
 
 #### Installation
 
@@ -36,7 +36,6 @@ Then see `docs/build/html/index.html` for the docs.
 34. mouse-specific intercepts
 36. SMC for sampling label sequence posterior
 37. Mouse-specific normalization options
-41. Group Lasso?
 42. Early stopping in `GridSearchCV`
 43. ``[b,f,r,r]`` vs ``[b,r,r,f]`` shapes
 45. automatic groups?
@@ -46,3 +45,4 @@ Then see `docs/build/html/index.html` for the docs.
 50. save GridSearchCV with validation scores for each parameter setting
 51. Load CSV label files
 52. Also save confusion matrices for ``GridSearchCV``
+53. Split file-related utilities off from utils.utils
