+++ lpne/__init__.py
@@ -5,7 +5,7 @@ Code for preprocessing and building factor models with local field potentials.
 
 """
 __date__ = "July 2021 - November 2022"
-__version__ = "0.1.9"
+__version__ = "0.1.10"
 try:
     with open(".git/logs/HEAD", "r") as fh:
         __commit__ = fh.read().split("\n")[-2]
@@ -20,6 +20,7 @@ from .models import (
     CpSae,
     GridSearchCV,
     DcsfaNmf,
+    get_model_class,
     get_reconstruction_stats,
     get_reconstruction_summary,
 )
@@ -55,15 +56,18 @@ from .preprocess.normalize import normalize_features, normalize_lfps
 from .preprocess.outlier_detection import mark_outliers
 
 from .utils.data import (
+    load_channel_map,
+    load_features,
+    load_features_and_labels,
+    load_labels,
     load_lfps,
     save_features,
-    load_features,
     save_labels,
-    load_labels,
-    load_features_and_labels,
 )
 
 from .utils.utils import (
+    get_all_fns,
+    infer_groups_from_fns,
     write_fake_labels,
     get_lfp_filenames,
     get_feature_filenames,
